backend/requirements.txt
pymongo
fastapi
cerberus 
passlib[bcrypt]
redis 
websockets
grpcio 
grpcio-tools
httpx